+++ elevatr/__init__.py
@@ -10,4 +10,4 @@ from ._version import __version__  # noqa F401
 from .get_elev_raster import get_elev_raster
 from .raster import Raster  # noqa F401
 
-__all__ = ["get_elev_raster"]
+__all__ = ["get_elev_raster", "Raster"]