+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Ask for user confirmation before downloading large files.
 - Add reprojection functionality to `get_elev_raster` function and `Raster` class.
+- Added the ability to access the resolution of the raster through the `Raster` class via the `raster.resolution`
 
 ## 0.2.0 - 2025-01-06
 ### Added