+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [Unreleased]
+## 0.3.0 - 2025-01-08
 ### Added
 - Ask for user confirmation before downloading large files.
 - Add reprojection functionality to `get_elev_raster` function and `Raster` class.